pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Siddhesh Parave",email = "siddeshparave16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "phonenumbers (>=9.0.7,<10.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
